refactor: applying localstorage. getting data from past instance
